Implement unified IO utils
